+++ app/src/main/java/sebgg/dev/workoutdiary/MainActivity.kt
@@ -4,12 +4,12 @@ import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.SavedStateViewModelFactory
-import sebgg.dev.workoutdiary.ui.helpers.SStateViewModel
-import sebgg.dev.workoutdiary.ui.history.HistoryFragment
+import sebgg.dev.workoutdiary.viewmodels.SStateViewModel
+import sebgg.dev.workoutdiary.ui.fragments.HistoryFragment
 import sebgg.dev.workoutdiary.ui.history.HistoryViewModel
 import sebgg.dev.workoutdiary.ui.history.HistoryViewModelFactory
-import sebgg.dev.workoutdiary.ui.main.MainFragment
-import sebgg.dev.workoutdiary.ui.workout.WorkoutFragment
+import sebgg.dev.workoutdiary.ui.fragments.MainFragment
+import sebgg.dev.workoutdiary.ui.fragments.WorkoutFragment
 import sebgg.dev.workoutdiary.ui.workout.WorkoutViewModel
 import sebgg.dev.workoutdiary.ui.workout.WorkoutViewModelFactory
 import java.util.*
