+++ app/src/main/java/sebgg/dev/workoutdiary/WorkoutApplication.kt
@@ -3,8 +3,8 @@ package sebgg.dev.workoutdiary
 import android.app.Application
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
-import sebgg.dev.workoutdiary.ui.database.ExerciseRepository
-import sebgg.dev.workoutdiary.ui.database.ExerciseRoomDatabase
+import sebgg.dev.workoutdiary.database.ExerciseRepository
+import sebgg.dev.workoutdiary.database.ExerciseRoomDatabase
 
 class WorkoutApplication: Application() {
     val applicationScope = CoroutineScope(SupervisorJob())
