+++ app/src/main/res/layout/main_fragment.xml
@@ -6,7 +6,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.main.MainFragment">
+        tools:context=".ui.fragments.MainFragment">
 
         <com.google.android.material.button.MaterialButton
             android:id="@+id/button_new_workout"
